High-level interface to specifying desirability functions

High-level interface to specifying desirability functions

Usage

desirability(..., .use_data = TRUE)

Arguments

...

using a goal function (see below) and the variable to be optimized. Other arguments should be specified as needed but must be named. Order of the arguments does not matter.

.use_data

A single logical to specify whether all translated desirability functions (such as d_max()) should enable use_data = TRUE to fill in any unspecified required arguments.

Value

An object of class "desirability_set" that can be used to make a set of desirability functions.

Details

The following set of nonexistent functions are used to specify an optimization goal:

maximize() (corresponding to d_max())

minimize() (d_min())

target() (d_target())

constrain() (d_box())

categories() (d_max())

For example, if you wanted to jointly maximize a regression model’s Rsquared while minimizing the RMSE, you could use

  desirability(
    minimize(rmse, scale = 3),
    maximize(rsq)
 )

Where the scale argument makes the desirability curve more stringent.