Desirability functions for selected data columns

These functions conducts the same desirability computations on multiple columns of a data set.

d_max_select(.data, ..., low, high, scale = 1, missing = 1)

Arguments

.data
A data frame or data frame extension (e.g. a tibble).

...
One or more unquoted expressions separated by commas (such as dplyr selector functions). Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables.

low
Single numeric values that define the active ranges of desirability.

high
Single numeric values that define the active ranges of desirability.

scale
A single numeric value to rescale the desirability function.

missing
A single numeric value on [0, 1] (or NA_real_) that defines how missing values in x are mapped to the desirability score.

Value

A tibble containing the select columns and their desirability scores. A prefix is added to the column names based on the type of desirability function (e.g., column x becomes column d_max_x when used with d_max_select()).

Examples

library(dplyr)

set.seed(1)
tibble(x = sort(runif(10)), y = sort(runif(10)), z = 1:10) %>%
  d_max_select(x, y, low = 0.5, high = 0.75)

#> # A tibble: 10 × 2
#>    d_max_x d_max_y
#>      <dbl>   <dbl>
#>  1   0       0    
#>  2   0       0    
#>  3   0       0    
#>  4   0       0    
#>  5   0.291   0    
#>  6   0.516   0.748
#>  7   0.643   0.870
#>  8   1       1    
#>  9   1       1    
#> 10   1       1